Create Calbuco_SupervisedBurnClassif
This script was designed to identify areas affected by a volcanic eruption, such as burned vegetation or ash deposits, using Landsat 8 satellite imagery, vegetation response, and classification based on training samples of the elements to be distinguished. In this case, the elements are "forest 1", "forest 2", "settlement", "roads", "burned vegetation", and "flooded valley" (due to pyroclastic flows).